@@ -2,52 +2,54 @@ import Register from './components/Register';
 import Login from './components/Login';
 import Home from './components/Home';
 import Layout from './components/Layout';
-import Editor from './components/Editor';
+// import Editor from './components/Editor';
 import Admin from './components/Admin';
 import Missing from './components/Missing';
 import Unauthorized from './components/Unauthorized';
 import Lounge from './components/Lounge';
 import LinkPage from './components/LinkPage';
 import RequireAuth from './components/RequireAuth';
-import PersistLogin from './components/PersistLogin';
 import { Routes, Route } from 'react-router-dom';
-
-const ROLES = {
-  'User': 2001,
-  'Editor': 1984,
-  'Admin': 5150
-}
+import Users from './components/Users';
+import PersistLogin from './components/PersistLogin';
+// import Events from './components/Events';
+// import AddEvent from './components/AddEvent';
+// import Test from './components/Test';
 
 function App() {
-
   return (
     <Routes>
       <Route path="/" element={<Layout />}>
         {/* public routes */}
+        {/* <Route path='test' element={<Test/>} /> */}
         <Route path="login" element={<Login />} />
         <Route path="register" element={<Register />} />
         <Route path="linkpage" element={<LinkPage />} />
         <Route path="unauthorized" element={<Unauthorized />} />
 
         {/* we want to protect these routes */}
-        <Route element={<PersistLogin />}>
-          <Route element={<RequireAuth allowedRoles={[ROLES.User]} />}>
+        <Route element={<PersistLogin/>}>
+          <Route element={<RequireAuth allowedRoles={['organisator']} />}>
+            
             <Route path="/" element={<Home />} />
+            {/* <Route path="editor" element={<Editor />} /> */}
+            {/* <Route path='events' element={<Events />} /> */}
           </Route>
 
-          <Route element={<RequireAuth allowedRoles={[ROLES.Editor]} />}>
-            <Route path="editor" element={<Editor />} />
-          </Route>
-
-
-          <Route element={<RequireAuth allowedRoles={[ROLES.Admin]} />}>
+          <Route element={<RequireAuth allowedRoles={['administrator']} />}>
             <Route path="admin" element={<Admin />} />
           </Route>
 
-          <Route element={<RequireAuth allowedRoles={[ROLES.Editor, ROLES.Admin]} />}>
-            <Route path="lounge" element={<Lounge />} />
+          <Route element={<RequireAuth allowedRoles={['user']} />}>
+            <Route path="users" element={<Users />} />
           </Route>
-        </Route>
+
+          <Route element={<RequireAuth allowedRoles={['organisator', 'administrator']} />}>
+            <Route path="lounge" element={<Lounge />} />
+            
+            {/* <Route path='addevent' element={<AddEvent/>} /> */}
+          </Route>  
+        </Route> 
 
         {/* catch all */}
         <Route path="*" element={<Missing />} />
