@@ -1,12 +1,20 @@
 import axios from 'axios';
-const BASE_URL = 'http://localhost:3500';
+const BASE_URL = 'https://localhost:7095'
 
 export default axios.create({
-    baseURL: BASE_URL
+    //baseURL: BASE_URL,
+    headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json'
+    },
+    //'Access-Control-Allow-Credentials': true
 });
 
 export const axiosPrivate = axios.create({
-    baseURL: BASE_URL,
-    headers: { 'Content-Type': 'application/json' },
-    withCredentials: true
+    //baseURL: BASE_URL,
+    headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json'
+    },
+    //'Access-Control-Allow-Credentials': true
 });
